OR1K: add support for non-hard mul, div and ror
expands mul, div and ror if the feature is not selected
hardware mul and div are on by default.
Rotate left is always expanded for now, hardware support
for this could be included in the rotate right by modifying
the rotate constant.
Also prepare for hardware support of conditional move (l.cmov),
by adding cmov feature.